requirements.txt
flask
python-docx
ollama
requests 